Moved cell implementation for sessions to a separate view to improve reusability
